Communication between Apple Watch and iPhone app
The aim of this post is explain how to implement the between Apple
Watch and iPhone app. After trying several methods: CoreData,
NSKeyedUnarchiver and NSUserDefaults. I have chosen NSUserDefault
because it was the most simplified one. I would recommend the other two
just in case that you wanted to share bigs amounts of data.